display duration needs bigger variable

+++ applications/settings.h
@@ -55,7 +55,7 @@ X (uint8_t,  disp_rot1,     $R, %i,    DISP_ROTATION1) \
 X (uint8_t,  disp_rot2,     $e, %i,    DISP_ROTATION2) \
 X (uint16_t, power_off_ms,  $p, %i,    DISP_POWER_ON_OFFTIME) \
 X (uint16_t, disp_beg_ms,   $f, %i,    DISP_OFF_TRIGGER_BEG_MS) \
-X (uint16_t, disp_dur_ms,   $D, %i,    DISP_OFF_TRIG_DURATION_MS) \
+X (uint32_t, disp_dur_ms,   $D, %i,    DISP_OFF_TRIG_DURATION_MS) \
 X (uint16_t, disp_on_ms,    $n, %i,    DISP_ON_TRIGGER_SPEED_MS) \
 X (float,    batt_imbalance,$i, %0.2f, BATTERY_MAX_IMBALANCE) \
 X (float,    b2Rratio,      $x, %0.2f, BATTERY2_SENSE_RATIO) \
